pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geocr-generator"
version = "1.0.0"
description = "A tool for generating GeoCroissant metadata from geospatial datasets"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "rasterio>=1.3.0",
    "numpy>=1.20.0",
]

[project.scripts]
geocr-generate = "geocr_generator.scripts.geocr_generate:main"

[tool.setuptools]
packages = ["geocr_generator", "geocr_generator.core", "geocr_generator.scripts"]
